Added a new leetcode-contest-ranking page
